+++ src/main/java/org/ahpuh/surf/user/service/UserService.java
@@ -1,23 +1,80 @@
 package org.ahpuh.surf.user.service;
 
+import lombok.RequiredArgsConstructor;
+import org.ahpuh.surf.follow.repository.FollowRepository;
+import org.ahpuh.surf.jwt.JwtAuthentication;
+import org.ahpuh.surf.jwt.JwtAuthenticationToken;
+import org.ahpuh.surf.user.converter.UserConverter;
 import org.ahpuh.surf.user.dto.UserDto;
 import org.ahpuh.surf.user.dto.UserJoinRequestDto;
 import org.ahpuh.surf.user.dto.UserLoginResponseDto;
 import org.ahpuh.surf.user.dto.UserUpdateRequestDto;
 import org.ahpuh.surf.user.entity.User;
+import org.ahpuh.surf.user.repository.UserRepository;
+import org.springframework.security.authentication.AuthenticationManager;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.crypto.password.PasswordEncoder;
+import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
-public interface UserService {
+import static org.ahpuh.surf.common.exception.EntityExceptionHandler.UserNotFound;
 
-    UserLoginResponseDto authenticate(final String email, final String password);
+@RequiredArgsConstructor
+@Transactional(readOnly = true)
+@Service
+public class UserService {
 
-    User login(final String email, final String password);
+    private final AuthenticationManager authenticationManager;
+    private final PasswordEncoder passwordEncoder;
 
-    Long join(final UserJoinRequestDto joinRequest);
+    private final UserRepository userRepository;
+    private final FollowRepository followRepository;
+    private final UserConverter userConverter;
 
-    UserDto findById(Long userId);
+    public UserLoginResponseDto authenticate(final String email, final String password) {
+        final JwtAuthenticationToken authToken = new JwtAuthenticationToken(email, password);
+        final Authentication resultToken = authenticationManager.authenticate(authToken);
+        final JwtAuthentication authentication = (JwtAuthentication) resultToken.getPrincipal();
+        final User user = (User) resultToken.getDetails();
+        return new UserLoginResponseDto(authentication.token, user.getUserId());
+    }
 
-    Long update(Long userId, UserUpdateRequestDto updateDto, String profilePhotoUrl);
+    public User login(final String email, final String password) {
+        final User user = userRepository.findByEmail(email)
+                .orElseThrow(() -> UserNotFound(email));
+        user.checkPassword(passwordEncoder, password);
+        return user;
+    }
 
-    void delete(Long userId);
+    @Transactional
+    public Long join(final UserJoinRequestDto joinRequest) {
+        if (userRepository.existsByEmail(joinRequest.getEmail())) {
+            throw new IllegalArgumentException(String.format("Email is duplicated. email=%s", joinRequest.getEmail()));
+        }
+        final User newUser = userRepository.save(userConverter.toEntity(joinRequest));
+        return newUser.getUserId();
+    }
 
+    public UserDto findById(final Long userId) {
+        final User user = userRepository.findById(userId)
+                .orElseThrow(() -> UserNotFound(userId));
+        final long followingCount = followRepository.countByUser(user);
+        final long followerCount = followRepository.countByFollowedUser(user);
+        return userConverter.toUserDto(user, followingCount, followerCount);
+    }
+
+    @Transactional
+    public Long update(final Long userId, final UserUpdateRequestDto updateDto, final String profilePhotoUrl) {
+        final User userEntity = userRepository.findById(userId)
+                .orElseThrow(() -> UserNotFound(userId));
+        userEntity.update(passwordEncoder, updateDto, profilePhotoUrl);
+        return userEntity.getUserId();
+    }
+
+    @Transactional
+    public void delete(final Long userId) {
+        final User userEntity = userRepository.findById(userId)
+                .orElseThrow(() -> UserNotFound(userId));
+        userRepository.delete(userEntity);
+    }
 }
